requirements.txt
flask
flask-cors
openai==0.28
python-docx
numpy
scikit-learn
